+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/dashboardnotifications/defendant/ClaimSettledDefendantDashboardNotificationHandler.java
@@ -55,8 +55,9 @@ public class ClaimSettledDefendantDashboardNotificationHandler extends Dashboard
         final String caseId = String.valueOf(caseData.getCcdCaseReference());
         boolean isLrQmEnabled = featureToggleService.isQueryManagementLRsEnabled();
 
-        if (!isLrQmEnabled || !featureToggleService.isGaForLipsEnabledAndLocationWhiteListed(caseData.getCaseManagementLocation().getBaseLocation())) {
-
+        if (!(isLrQmEnabled &&
+            caseData.getPreviousCCDState() == null &&
+            !featureToggleService.isGaForLipsEnabledAndLocationWhiteListed(caseData.getCaseManagementLocation().getBaseLocation()))) {
             dashboardNotificationService.deleteByReferenceAndCitizenRole(
                 caseId,
                 "DEFENDANT"
